Allow for corrections where edges make a cycle
* If edges aren't in a cycle, it has no impact. Allowing us to
  train in batches with the same logic as cycles
