+++ fe/free_energy.py
@@ -1,3 +1,4 @@
+from typing import List
 from jax.config import config; config.update("jax_enable_x64", True)
 
 import jax
@@ -9,6 +10,8 @@ from timemachine.lib import potentials, custom_ops
 from timemachine.lib import LangevinIntegrator
 
 from ff.handlers import openmm_deserializer
+
+from rdkit.Chem import MolToSmiles
 
 def get_romol_conf(mol):
     """Coordinates of mol's 0th conformer, in nanometers"""
@@ -219,6 +222,34 @@ class RelativeFreeEnergy(BaseFreeEnergy):
         combined_coords = np.concatenate([host_coords, np.mean(self.top.interpolate_params(ligand_coords_a, ligand_coords_b), axis=0)])
 
         return final_potentials, final_params, combined_masses, combined_coords
+
+
+class RBFETransformIndex:
+    """Builds an index of relative free energy transformations to use
+    with construct_mle_layer
+    """
+
+    def __len__(self):
+        return len(self._indices)
+
+    def __init__(self):
+        self._indices = {}
+
+    def build(self, refs: List[RelativeFreeEnergy]):
+        for ref in refs:
+            self.get_transform_indices(ref)
+
+    def get_transform_indices(self, ref: RelativeFreeEnergy) -> List[int]:
+        return self.get_mol_idx(ref.mol_a), self.get_mol_idx(ref.mol_b)
+
+    def get_mol_idx(self, mol):
+        hashed = self._mol_hash(mol)
+        if hashed not in self._indices:
+            self._indices[hashed] = len(self._indices)
+        return self._indices[hashed]
+
+    def _mol_hash(self, mol):
+        return MolToSmiles(mol)
 
 
 def construct_lambda_schedule(num_windows):
